ROX-35326: Address PR review feedback
Split PrometheusMetricsCard.tsx into PrometheusMetricsTabbedCard.tsx and
PrometheusMetricsTabbedForm.tsx to match file-per-export convention.
Remove PrivateConfigPrometheusMetricsDetails wrapper. Replace PF utility
class with Flex component. Show metric count badge when count is 0.
Rename pcfg prop to privateConfig.

Partially generated by AI.
