--- ui/apps/platform/src/Containers/SystemConfig/Details/PrivateConfigPrometheusMetricsDetails.tsx
@@ -1,16 +1,0 @@
-import type { ReactElement } from 'react';
-
-import type { PrivateConfig } from 'types/config.proto';
-import { PrometheusMetricsTabbedCard } from './components/PrometheusMetricsCard';
-
-export type PrivateConfigPrometheusMetricsDetailsProps = {
-    privateConfig: PrivateConfig;
-};
-
-const PrivateConfigPrometheusMetricsDetails = ({
-    privateConfig,
-}: PrivateConfigPrometheusMetricsDetailsProps): ReactElement => {
-    return <PrometheusMetricsTabbedCard privateConfig={privateConfig} />;
-};
-
-export default PrivateConfigPrometheusMetricsDetails;
