+++ ui/apps/platform/src/Containers/SystemConfig/Details/components/PrometheusMetricsTabbedCard.tsx
@@ -10,16 +10,12 @@ import {
     DescriptionListTerm,
     EmptyState,
     EmptyStateBody,
-    FormGroup,
-    FormSection,
-    Grid,
-    GridItem,
+    Flex,
     Label,
     LabelGroup,
     Tab,
     TabTitleText,
     Tabs,
-    TextInput,
 } from '@patternfly/react-core';
 import { MinusIcon, PlusIcon } from '@patternfly/react-icons';
 import { Table, Tbody, Td, Th, Thead, Tr } from '@patternfly/react-table';
@@ -32,14 +28,14 @@ import type {
     PrometheusMetricsLabels,
 } from 'types/config.proto';
 
-const metricPrefixes = {
+export const metricPrefixes = {
     imageVulnerabilities: 'rox_central_image_vuln_',
     nodeVulnerabilities: 'rox_central_node_vuln_',
     policyViolations: 'rox_central_policy_violation_',
     administrativeEvents: 'rox_central_admin_event_',
 };
 
-const predefinedMetrics: Record<
+export const predefinedMetrics: Record<
     PrometheusMetricsCategory,
     Record<string, PrometheusMetricsLabels>
 > = {
@@ -99,11 +95,11 @@ const predefinedMetrics: Record<
     },
 };
 
-export type PrometheusMetricsLabelGroupProps = {
+function PrometheusMetricsLabelGroup({
+    labels,
+}: {
     labels: PrometheusMetricsLabels;
-};
-
-function PrometheusMetricsLabelGroup({ labels }: PrometheusMetricsLabelGroupProps): ReactElement {
+}): ReactElement {
     return (
         <LabelGroup isCompact numLabels={Infinity}>
             {labels.labels.map((label) => {
@@ -117,11 +113,11 @@ function PrometheusMetricsLabelGroup({ labels }: PrometheusMetricsLabelGroupProp
     );
 }
 
-export type PrometheusMetricsFilterGroupProps = {
+function PrometheusMetricsFilterGroup({
+    labels,
+}: {
     labels: PrometheusMetricsLabels;
-};
-
-function PrometheusMetricsFilterGroup({ labels }: PrometheusMetricsFilterGroupProps): ReactElement {
+}): ReactElement {
     const includeEntries = Object.entries(labels.includeFilters ?? {}).sort(([a], [b]) =>
         a.localeCompare(b)
     );
@@ -148,7 +144,62 @@ function PrometheusMetricsFilterGroup({ labels }: PrometheusMetricsFilterGroupPr
     );
 }
 
-export type PrometheusMetricsPredefinedMetricTableRowProps = {
+function recordsMatch(a: Record<string, string>, b: Record<string, string>): boolean {
+    return (
+        Object.keys(a).length === Object.keys(b).length &&
+        Object.entries(a).every(([key, val]) => b[key] === val)
+    );
+}
+
+function elementsMatch(a: string[], b: string[]): boolean {
+    return a.length === b.length && a.every((value) => b.includes(value));
+}
+
+export function hasMetric(
+    descriptors: Record<string, PrometheusMetricsLabels> | undefined,
+    metric: string,
+    labels: PrometheusMetricsLabels
+): boolean {
+    const base = {
+        labels: descriptors?.[metric]?.labels ?? [],
+        includeFilters: descriptors?.[metric]?.includeFilters ?? {},
+        excludeFilters: descriptors?.[metric]?.excludeFilters ?? {},
+    };
+    const given = {
+        labels: labels.labels,
+        includeFilters: labels.includeFilters ?? {},
+        excludeFilters: labels.excludeFilters ?? {},
+    };
+    return (
+        elementsMatch(base.labels, given.labels) &&
+        recordsMatch(base.includeFilters, given.includeFilters) &&
+        recordsMatch(base.excludeFilters, given.excludeFilters)
+    );
+}
+
+function hasFiltersInLabels(labels: PrometheusMetricsLabels): boolean {
+    return (
+        Object.keys(labels.includeFilters ?? {}).length > 0 ||
+        Object.keys(labels.excludeFilters ?? {}).length > 0
+    );
+}
+
+function metricsHaveFilters(
+    descriptors: Record<string, PrometheusMetricsLabels> | undefined,
+    category: PrometheusMetricsCategory,
+    editMode: boolean
+): boolean {
+    const hasFiltersInPredefined = editMode
+        ? Object.values(predefinedMetrics[category]).some(hasFiltersInLabels)
+        : Object.entries(predefinedMetrics[category]).some(
+              ([metric, labels]) =>
+                  hasMetric(descriptors, metric, labels) && hasFiltersInLabels(labels)
+          );
+    const hasFiltersInDescriptors = Object.values(descriptors ?? {}).some(hasFiltersInLabels);
+    return hasFiltersInPredefined || hasFiltersInDescriptors;
+}
+
+type PrometheusMetricsPredefinedMetricTableRowProps = {
     category: PrometheusMetricsCategory;
     enabled: boolean;
     metric: string;
@@ -206,69 +257,7 @@ function PrometheusMetricsPredefinedMetricTableRow({
     );
 }
 
-// recordsMatch returns true if two maps are equal.
-function recordsMatch(a: Record<string, string>, b: Record<string, string>): boolean {
-    return (
-        Object.keys(a).length === Object.keys(b).length &&
-        Object.entries(a).every(([key, val]) => b[key] === val)
-    );
-}
-
-// elementsMatch returns true if two arrays have same elements ignoring order.
-function elementsMatch(a: string[], b: string[]): boolean {
-    return a.length === b.length && a.every((value) => b.includes(value));
-}
-
-// hasMetric checks if the descriptors contain the given metric by looking at
-// the metric name and the labels (ignoring the order).
-function hasMetric(
-    descriptors: Record<string, PrometheusMetricsLabels> | undefined,
-    metric: string,
-    labels: PrometheusMetricsLabels
-): boolean {
-    const base = {
-        labels: descriptors?.[metric]?.labels ?? [],
-        includeFilters: descriptors?.[metric]?.includeFilters ?? {},
-        excludeFilters: descriptors?.[metric]?.excludeFilters ?? {},
-    };
-    const given = {
-        labels: labels.labels,
-        includeFilters: labels.includeFilters ?? {},
-        excludeFilters: labels.excludeFilters ?? {},
-    };
-    return (
-        elementsMatch(base.labels, given.labels) &&
-        recordsMatch(base.includeFilters, given.includeFilters) &&
-        recordsMatch(base.excludeFilters, given.excludeFilters)
-    );
-}
-
-function hasFiltersInLabels(labels: PrometheusMetricsLabels): boolean {
-    return (
-        Object.keys(labels.includeFilters ?? {}).length > 0 ||
-        Object.keys(labels.excludeFilters ?? {}).length > 0
-    );
-}
-
-function metricsHaveFilters(
-    descriptors: Record<string, PrometheusMetricsLabels> | undefined,
-    category: PrometheusMetricsCategory,
-    editMode: boolean
-): boolean {
-    // In edit mode, check all predefined metrics.
-    // In view mode, only check enabled metrics (those in descriptors).
-    const hasFiltersInPredefined = editMode
-        ? Object.values(predefinedMetrics[category]).some(hasFiltersInLabels)
-        : Object.entries(predefinedMetrics[category]).some(
-              ([metric, labels]) =>
-                  hasMetric(descriptors, metric, labels) && hasFiltersInLabels(labels)
-          );
-    const hasFiltersInDescriptors = Object.values(descriptors ?? {}).some(hasFiltersInLabels);
-    const showFilters = hasFiltersInPredefined || hasFiltersInDescriptors;
-    return showFilters;
-}
-
-type PrometheusMetricsTableProps = {
+export type PrometheusMetricsTableProps = {
     descriptors: Record<string, PrometheusMetricsLabels> | undefined;
     category: PrometheusMetricsCategory;
     onCustomChange:
@@ -276,7 +265,7 @@ type PrometheusMetricsTableProps = {
         | undefined;
 };
 
-function PrometheusMetricsTable({
+export function PrometheusMetricsTable({
     descriptors,
     category,
     onCustomChange,
@@ -297,10 +286,6 @@ function PrometheusMetricsTable({
             <Tbody>
                 {Object.entries(predefinedMetrics[category]).map(
                     ([predefinedMetric, originalLabels], rowIndex) => {
-                        // In view mode show only enabled predefined metrics.
-                        // In edit mode show all predefined metrics unless they're
-                        // overridden.
-
                         const isEnabledOriginal = hasMetric(
                             descriptors,
                             predefinedMetric,
@@ -324,7 +309,6 @@ function PrometheusMetricsTable({
                     }
                 )}
                 {Object.entries(descriptors ?? {}).map(([metric, labels]) => {
-                    // Predefined are rendered above.
                     if (hasMetric(predefinedMetrics[category], metric, labels)) {
                         return null;
                     }
@@ -360,33 +344,28 @@ function PrometheusMetricsTable({
     );
 }
 
-const categoryTitles: Record<PrometheusMetricsCategory, string> = {
+export const categoryTitles: Record<PrometheusMetricsCategory, string> = {
     imageVulnerabilities: 'Image vulnerabilities',
     nodeVulnerabilities: 'Node vulnerabilities',
     policyViolations: 'Policy violations',
     administrativeEvents: 'Administrative events',
 };
 
-function getMetricCount(descriptors: Record<string, PrometheusMetricsLabels> | undefined): number {
+export function getMetricCount(
+    descriptors: Record<string, PrometheusMetricsLabels> | undefined
+): number {
     return descriptors ? Object.keys(descriptors).length : 0;
 }
 
-function MetricCountBadge({ count }: { count: number }): ReactElement | null {
-    if (count === 0) {
-        return null;
-    }
-    return (
-        <Badge isRead className="pf-v6-u-ml-sm">
-            {count}
-        </Badge>
-    );
+export function MetricCountBadge({ count }: { count: number }): ReactElement {
+    return <Badge isRead>{count}</Badge>;
 }
 
 export type PrometheusMetricsTabbedCardProps = {
     privateConfig: PrivateConfig;
 };
 
-export function PrometheusMetricsTabbedCard({
+export default function PrometheusMetricsTabbedCard({
     privateConfig,
 }: PrometheusMetricsTabbedCardProps): ReactElement {
     const categories = Object.keys(categoryTitles) as PrometheusMetricsCategory[];
@@ -411,8 +390,10 @@ export function PrometheusMetricsTabbedCard({
                             eventKey={category}
                             title={
                                 <TabTitleText>
-                                    {categoryTitles[category]}
-                                    <MetricCountBadge count={metricCount} />
+                                    <Flex gap={{ default: 'gapSm' }}>
+                                        {categoryTitles[category]}
+                                        <MetricCountBadge count={metricCount} />
+                                    </Flex>
                                 </TabTitleText>
                             }
                         >
@@ -455,105 +436,6 @@ export function PrometheusMetricsTabbedCard({
                                         </EmptyStateBody>
                                     </EmptyState>
                                 )}
-                            </CardBody>
-                        </Tab>
-                    );
-                })}
-            </Tabs>
-        </Card>
-    );
-}
-
-type PrometheusMetricsPeriodFormProps = {
-    pcfg: PrivateConfig;
-    category: PrometheusMetricsCategory;
-    onChange: (value, event) => Promise<void> | Promise<FormikErrors<FormikValues>>;
-};
-
-function PrometheusMetricsPeriodForm({
-    pcfg,
-    category,
-    onChange,
-}: PrometheusMetricsPeriodFormProps): ReactElement {
-    return (
-        <FormGroup
-            label="Gathering period in minutes (set to 0 to disable)"
-            isRequired
-            fieldId={`privateConfig.metrics.${category}.gatheringPeriodMinutes`}
-        >
-            <TextInput
-                isRequired
-                type="number"
-                id={`privateConfig.metrics.${category}.gatheringPeriodMinutes`}
-                name={`privateConfig.metrics.${category}.gatheringPeriodMinutes`}
-                value={pcfg?.metrics?.[category]?.gatheringPeriodMinutes}
-                onChange={(event, value) => onChange(value, event)}
-                min={0}
-            />
-        </FormGroup>
-    );
-}
-
-export type PrometheusMetricsTabbedFormProps = {
-    pcfg: PrivateConfig;
-    onChange: (value, event) => Promise<void> | Promise<FormikErrors<FormikValues>>;
-    onCustomChange?: (
-        value: unknown,
-        id: string
-    ) => Promise<void> | Promise<FormikErrors<FormikValues>>;
-};
-
-export function PrometheusMetricsTabbedForm({
-    pcfg,
-    onChange,
-    onCustomChange,
-}: PrometheusMetricsTabbedFormProps): ReactElement {
-    const categories = Object.keys(categoryTitles) as PrometheusMetricsCategory[];
-    const [activeTab, setActiveTab] = useState<PrometheusMetricsCategory>(categories[0]);
-
-    return (
-        <Card data-testid="prometheus-metrics-config">
-            <Tabs
-                activeKey={activeTab}
-                onSelect={(_event, tabKey) => setActiveTab(tabKey as PrometheusMetricsCategory)}
-            >
-                {categories.map((category) => {
-                    const metricCount = getMetricCount(pcfg?.metrics?.[category]?.descriptors);
-
-                    return (
-                        <Tab
-                            key={category}
-                            eventKey={category}
-                            title={
-                                <TabTitleText>
-                                    {categoryTitles[category]}
-                                    <MetricCountBadge count={metricCount} />
-                                </TabTitleText>
-                            }
-                        >
-                            <CardBody>
-                                <FormSection>
-                                    <Grid hasGutter>
-                                        <GridItem md={12}>
-                                            <PrometheusMetricsPeriodForm
-                                                pcfg={pcfg}
-                                                category={category}
-                                                onChange={onChange}
-                                            />
-                                        </GridItem>
-                                        <GridItem md={12}>
-                                            <FormGroup label="Metrics configuration" role="group">
-                                                <PrometheusMetricsTable
-                                                    descriptors={
-                                                        pcfg?.metrics?.[category]?.descriptors
-                                                    }
-                                                    category={category}
-                                                    onCustomChange={onCustomChange}
-                                                />
-                                            </FormGroup>
-                                        </GridItem>
-                                    </Grid>
-                                </FormSection>
                             </CardBody>
                         </Tab>
                     );
